Move proxy to middleware
Also change routes to use Express for the whitelist filtering
this allows us to use Express path syntax
