+++ src/middleware/api-proxy.js
@@ -1,0 +1,38 @@
+const proxy = require('http-proxy-middleware')
+
+const config = require('../config/')
+
+const API_PROXY_PATH = '/api-proxy'
+const WHITELIST = [
+  '/v4/company-list',
+  '/v4/company-list/:id/item',
+  '/v4/search/export-country-history',
+  '/v4/company-referral/',
+  '/v4/company-referral/:id',
+]
+
+module.exports = (app) => {
+  app.use(
+    WHITELIST.map((apiPath) => API_PROXY_PATH + apiPath),
+    proxy('/', {
+      changeOrigin: true,
+      target: config.apiRoot,
+      pathRewrite: {
+        ['^' + API_PROXY_PATH]: '',
+      },
+      onProxyReq: (proxyReq, req) => {
+        proxyReq.setHeader('authorization', `Bearer ${req.session.token}`)
+        // this is required to be able to handle POST requests and avoid a conflict with bodyParser
+        // issue here -> https://github.com/chimurai/http-proxy-middleware/issues/320
+        if (req.body) {
+          let bodyData = JSON.stringify(req.body)
+          // incase if content-type is application/x-www-form-urlencoded -> we need to change to application/json
+          proxyReq.setHeader('Content-Type', 'application/json')
+          proxyReq.setHeader('Content-Length', Buffer.byteLength(bodyData))
+          // stream the content
+          proxyReq.write(bodyData)
+        }
+      },
+    })
+  )
+}
